# api/openapi.yaml
openapi: 3.0.3
info:
  title: Licenses Admin API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /variants/{id}:
    get:
      summary: Get variant by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVariant'
        '404':
          description: Not found
    put:
      summary: Update variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVariantUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVariant'
        '404':
          description: Not found
    delete:
      summary: Delete variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: No content
  /variants/{id}/availability:
    get:
      summary: Get variant availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantAvailability'
...
components:
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    ProductVariant:
      type: object
      required: [id, productId, key, name]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        productId: { $ref: '#/components/schemas/UUID' }
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
...
    ProductVariantUpdateRequest:
      type: object
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
...
    VariantAvailability:
      type: object
      required: [capacity, assignedActiveCount, available]
      properties:
        capacity: { type: integer, nullable: true }
        assignedActiveCount: { type: integer }
        available: { type: boolean }
...