# target/generated-sources/openapi/src/main/java/io/meimberg/licenses/web/api/VariantsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.meimberg.licenses.web.api;

import io.meimberg.licenses.web.dto.ProductVariant;
import io.meimberg.licenses.web.dto.ProductVariantUpdateRequest;
import java.util.UUID;
import io.meimberg.licenses.web.dto.VariantAvailability;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T15:50:51.445317955+01:00[Europe/Berlin]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "variants", description = "the variants API")
public interface VariantsApi {

    /**
     * GET /variants/{id}/availability : Get variant availability
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "variantsIdAvailabilityGet",
        summary = "Get variant availability",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = VariantAvailability.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/variants/{id}/availability",
        produces = { "application/json" }
    )
    
    ResponseEntity<VariantAvailability> variantsIdAvailabilityGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * DELETE /variants/{id} : Delete variant
     *
     * @param id  (required)
     * @return No content (status code 204)
     */
    @Operation(
        operationId = "variantsIdDelete",
        summary = "Delete variant",
        responses = {
            @ApiResponse(responseCode = "204", description = "No content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/variants/{id}"
    )
    
    ResponseEntity<Void> variantsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /variants/{id} : Get variant by id
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "variantsIdGet",
        summary = "Get variant by id",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductVariant.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/variants/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ProductVariant> variantsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * PUT /variants/{id} : Update variant
     *
     * @param id  (required)
     * @param productVariantUpdateRequest  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "variantsIdPut",
        summary = "Update variant",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductVariant.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/variants/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductVariant> variantsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "ProductVariantUpdateRequest", description = "", required = true) @Valid @RequestBody ProductVariantUpdateRequest productVariantUpdateRequest
    );

}